Buildable

+++ Source/RW_Patches/Pawn_PathFollower_Patch.cs
@@ -14,9 +14,9 @@ namespace RimThreaded.RW_Patches
             RimThreadedHarmony.Prefix(original, patched, nameof(CostToMoveIntoCell), new Type[] { typeof(Pawn), typeof(IntVec3) });
         }
 
-        public static bool CostToMoveIntoCell(ref int __result, Pawn pawn, IntVec3 c)
+        public static bool CostToMoveIntoCell(ref float __result, Pawn pawn, IntVec3 c)
         {
-            int a = (c.x == pawn.Position.x || c.z == pawn.Position.z ? pawn.TicksPerMoveCardinal : pawn.TicksPerMoveDiagonal) + pawn.Map.pathing.For(pawn).pathGrid.CalculatedCostAt(c, false, pawn.Position);
+            float a = (c.x == pawn.Position.x || c.z == pawn.Position.z ? pawn.TicksPerMoveCardinal : pawn.TicksPerMoveDiagonal) + pawn.Map.pathing.For(pawn).pathGrid.CalculatedCostAt(c, false, pawn.Position);
             Building edifice = c.GetEdifice(pawn.Map);
             if (edifice != null)
                 a += edifice.PathWalkCostFor(pawn);
@@ -27,7 +27,7 @@ namespace RimThreaded.RW_Patches
                 Pawn locomotionUrgencySameAs = pawn?.jobs?.curDriver?.locomotionUrgencySameAs; //changed
                 if (locomotionUrgencySameAs != null && locomotionUrgencySameAs != pawn && locomotionUrgencySameAs.Spawned)
                 {
-                    int moveIntoCell = Pawn_PathFollower.CostToMoveIntoCell(locomotionUrgencySameAs, c);
+                    float moveIntoCell = Pawn_PathFollower.CostToMoveIntoCell(locomotionUrgencySameAs, c);
                     if (a < moveIntoCell)
                         a = moveIntoCell;
                 }
